% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price.R
\name{price}
\alias{price}
\title{Extract price data history}
\usage{
price(from, to = Sys.Date())
}
\arguments{
\item{from}{start period data extraction format yyyy-mm-dd}

\item{to}{end period data extraction format yyyy-mm-dd}
}
\description{
Get the price of rice within period.
}
\examples{
price(from = "2023-07-13", to = "2023-07-31")



}
