% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{stock}
\alias{stock}
\title{Extract rice stock data history}
\usage{
stock(m = NULL)
}
\arguments{
\item{m}{integer indicate month. January = 1, and December = 12. You can omit `m` to get all stock data current year.}
}
\description{
Get the price of rice in a month current year.
}
\examples{
# to extract stock history on January current year (i.e 2023)
stock(m = 1)

# to extract stock history on whole current year (i.e 2023)
stock()


}
